Update the GroupDocs.Metadata dependency

+++ Examples/src/main/java/com/groupdocs/metadata/examples/advanced_usage/managing_metadata_for_specific_formats/image/jpeg/maker_note/MakerNoteReadAllTags.java
@@ -21,7 +21,7 @@ public class MakerNoteReadAllTags {
                 for (TiffTag tag : root.getMakerNotePackage().toList()) {
 
                     // Please note that tag ids used by camera manufacturers may intersect with the ids defined in the TIFF/EXIF specification
-                    System.out.println(String.format("%s = %s", (int) tag.getTagID(), tag.getValue()));
+                    System.out.println(String.format("%s = %s", tag.getTagID(), tag.getValue()));
                 }
             }
         }
